finagle/finagle-core: Introduce SLOStatsFilter for recording SLO violations
Problem

Currently, server-side stats expose success rate and latency metrics, but there is no
way to know how often a target latency is being violated.

Solution

Introduce SLOStatsFilter (still in development, not yet added to the default server stack)
that can be configured with a target latency. This filter records both latency violations and
response failure violations to provide a picture of the server's SLO compliance.

Differential Revision: https://phabricator.twitter.biz/D1191333