finagle-core: Halve Netty tasks in Netty4 pipeline client
Problem
The pipelining client is used for the Memcache client. Extra tasks are
generated for the Netty event loop during message sending and receiving.

Solution
When receiving a response from the server, we process the response in
the current thread if it is a Netty event loop thread. When sending a message,
we do not hand off the message to a Netty thread inside Netty4PushChannelHandle,
if the sending is already performed within the Netty event loop.

Differential Revision: https://phabricator.twitter.biz/D1193718