finagle/finagle-core: Add retry context to retried requests
Problem

Currently, it's hard to attribute a surge in request volume and failures with a
retry storm -- it's a positive feedback loop. In a system with a complex call
chain, it can be difficult to know where to look for client-side retries, and these
still may not accurately reflect what the server ultimately sees due to multiplicative
retries at different levels.

Solution

Introduce a new context, `c.t.f.Retry`, to signal that a request is a retry. We can then
add server-side stats to be able to determine how many retries we receive (and look
at the against total request volume).

Differential Revision: https://phabricator.twitter.biz/D1228127